regression test suite passes
