+++ saml-proxy/package.json
@@ -123,12 +123,12 @@
     "eslint-config-prettier": "^8.1.0",
     "eslint-plugin-jest": "^24.2.0",
     "eslint-plugin-prettier": "^3.3.1",
-    "saml-attacks": "github:lukemarkwordtlibertyits/saml-attacks",
-    "saml-encoder-decoder-js": "^1.0.1",
     "jest": "^26.6.3",
     "nodemon": "^2.0.7",
     "prettier": "2.2.1",
     "puppeteer": "^10.0.0",
+    "saml-attacks": "github:lukemarkwordtlibertyits/saml-attacks",
+    "saml-encoder-decoder-js": "^1.0.1",
     "source-map-support": "^0.5.19",
     "ts-jest": "^26.5.3",
     "ts-node": "^9.1.1",
