--- saml-proxy/test/regression_tests/page-assertions.js
@@ -1,87 +1,0 @@
-require("jest");
-
-const isError = async (page, errorMessage) => {
-  const heading = await page.$eval(
-    ".usa-alert-body > .usa-alert-heading",
-    (elem) => elem.innerText
-  );
-  expect(heading).toEqual("Error");
-
-  const errorText = await page.$eval(
-    ".usa-alert-body > p",
-    (elem) => elem.innerText
-  );
-  expect(errorText).toEqual(errorMessage);
-
-  const errorIdHeadding = await page.$eval(
-    ".usa-alert-body > .usa-alert-heading:nth-of-type(2)",
-    (elem) => elem.innerText
-  );
-  expect(errorIdHeadding).toEqual("Error ID");
-
-  const errorId = await page.$eval(
-    ".usa-alert-body > pre",
-    (elem) => elem.innerText
-  );
-  expect(errorId);
-};
-
-const isSensitiveError = async (page) => {
-  const heading = await page.$eval(
-    ".usa-alert-body > .usa-alert-heading",
-    (elem) => elem.innerText
-  );
-  expect(heading).toEqual("Error");
-
-  const errorText = await page.$eval(
-    ".usa-alert-body > p",
-    (elem) => elem.innerText
-  );
-  expect(errorText).toEqual("Your request could not be processed.");
-
-  const errorIdHeadding = await page.$eval(
-    ".usa-alert-body > .usa-alert-heading:nth-of-type(2)",
-    (elem) => elem.innerText
-  );
-  expect(errorIdHeadding).toEqual("Error ID");
-
-  const errorId = await page.$eval(
-    ".usa-alert-body > pre",
-    (elem) => elem.innerText
-  );
-  expect(errorId);
-};
-
-const isIcnError = async (page) => {
-  const heading = await page.$eval(
-    ".saml-proxy > div > h2",
-    (elem) => elem.innerText
-  );
-  expect(heading).toEqual("Matching Error");
-
-  const alertHeading = await page.$eval(
-    ".usa-alert-heading",
-    (elem) => elem.innerText
-  );
-  expect(alertHeading).toEqual(
-    "We need to verify your identity before giving you access to your information"
-  );
-
-  const errorIdHeadding = await page.$eval(
-    ".usa-alert-body > .usa-alert-heading:nth-of-type(2)",
-    (elem) => elem.innerText
-  );
-  expect(errorIdHeadding).toEqual("Error ID");
-
-  const errorId = await page.$eval(
-    ".usa-alert-body > pre",
-    (elem) => elem.innerText
-  );
-  expect(errorId);
-};
-
-module.exports = {
-  isError,
-  isSensitiveError,
-  isIcnError,
-};
