+++ saml-proxy/test/regression_tests/saml-proxy-regression.test.js
@@ -1,7 +1,6 @@
 require("jest");
 const { v4: uuidv4 } = require("uuid");
 const puppeteer = require("puppeteer");
-const { isSensitiveError, isIcnError } = require("./page-assertions");
 const qs = require("querystring");
 const { ModifyAttack } = require("saml-attacks");
 const SAML = require("saml-encoder-decoder-js");
@@ -19,146 +18,145 @@ const defaultScope = [
 ];
 
 const authorization_url = "https://sandbox-api.va.gov/oauth2";
+const saml_proxy_url = process.env.SAML_PROXY_URL;
 const redirect_uri = "https://app/after-auth";
 
-jest.setTimeout(30000);
-
-test("Happy Path", async () => {
-  const browser = await puppeteer.launch(launchArgs);
-  const page = await browser.newPage();
-  await page.setDefaultNavigationTimeout(30000);
-  await requestToken(page);
-
-  let code = await login(
-    page,
-    "va.api.user+idme.001@gmail.com",
-    "Password1234!",
-    true
-  );
-  expect(code).not.toBeNull();
-  await browser.close();
-});
-
-test("ICN Error", async () => {
-  const browser = await puppeteer.launch(launchArgs);
-  const page = await browser.newPage();
-  await requestToken(page);
-
-  await login(page, "va.api.user+idme.043@gmail.com", "Password1234!");
-
-  await page.waitForRequest((request) => {
-    return request.url().includes("/samlproxy/sp/saml/sso");
+describe("Regression tests", () => {
+  jest.setTimeout(30000);
+  let browser;
+  beforeEach(async () => {
+    browser = await puppeteer.launch(launchArgs);
   });
 
-  await page.waitForNavigation({
-    waitUntil: "networkidle0",
+  afterEach(async () => {
+    await browser.close();
   });
 
-  await isIcnError(page);
-  await browser.close();
-});
+  test("Happy Path", async () => {
+    const page = await browser.newPage();
+    await page.setDefaultNavigationTimeout(30000);
+    await requestToken(page);
 
-test("Replay", async () => {
-  const browser = await puppeteer.launch(launchArgs);
-  const page = await browser.newPage();
-
-  await requestToken(page);
-  await authentication(page);
-
-  let post_data;
-  await page.waitForRequest((request) => {
-    if (request.url().includes("/samlproxy/sp/saml/sso")) {
-      post_data = request.postData();
-      return true;
-    }
-    return false;
+    let code = await login(
+      page,
+      "va.api.user+idme.001@gmail.com",
+      "Password1234!",
+      true
+    );
+    expect(code).not.toBeNull();
   });
 
-  await page.waitForRequest((request) =>
-    request.url().includes("https://app/after-auth")
-  );
-  let anotherPage = await browser.newPage();
-  await anotherPage.setRequestInterception(true);
-  anotherPage.on("request", async (request) => {
-    if (
-      request.url().includes("/samlproxy/sp/saml/sso") &&
-      request.method() == "GET"
-    ) {
-      await request.continue({
-        method: "POST",
-        postData: post_data,
-        headers: {
-          ...request.headers(),
-          "Content-Type": "application/x-www-form-urlencoded",
-        },
-      });
-    } else {
-      await request.continue();
-    }
-  });
-  await anotherPage.goto(`${authorization_url}/samlproxy/sp/saml/sso`);
+  test("ICN Error", async () => {
+    const page = await browser.newPage();
+    await requestToken(page);
 
-  // assertion
-  await isSensitiveError(anotherPage);
-  await browser.close();
-});
+    await login(page, "va.api.user+idme.043@gmail.com", "Password1234!");
 
-test("modify", async () => {
-  const browser = await puppeteer.launch(launchArgs);
-  const page = await browser.newPage();
-  await requestToken(page);
-  await authentication(page, "va.api.user+idme.001@gmail.com", true);
+    await page.waitForRequest((request) => {
+      return request.url().includes("/samlproxy/sp/saml/sso");
+    });
 
-  page.on("request", async (request) => {
-    if (
-      request.url().includes("/samlproxy/sp/saml/sso") &&
-      request.method() == "POST"
-    ) {
-      let post_data = decode(request);
-      post_data.SAMLResponse = ModifyAttack(
-        post_data.SAMLResponse,
-        "uuid",
-        "modify"
-      );
+    await page.waitForNavigation({
+      waitUntil: "networkidle0",
+    });
 
-      await request.continue({
-        method: "POST",
-        postData: encode(post_data),
-        headers: {
-          ...request.headers(),
-          "Content-Type": "application/x-www-form-urlencoded",
-        },
-      });
-    } else {
-      await request.continue();
-    }
+    await isError(
+      page,
+      "Some VA.gov tools aren't working right now",
+      "We're sorry. Something went wrong on our end while looking up your account. You may not be able to connect to your VA records until we can figure out what's wrong."
+    );
   });
 
-  await page.waitForSelector(".usa-alert-error");
+  test("Replay", async () => {
+    const page = await browser.newPage();
 
-  await isSensitiveError(page);
+    await requestToken(page);
+    await authentication(page);
 
-  await browser.close();
-});
+    let post_data;
+    await page.waitForRequest((request) => {
+      if (request.url().includes("/samlproxy/sp/saml/sso")) {
+        post_data = request.postData();
+        return true;
+      }
+      return false;
+    });
 
-test("Empty SSO", async () => {
-  const browser = await puppeteer.launch(launchArgs);
-  const page = await browser.newPage();
+    await page.waitForRequest((request) =>
+      request.url().includes("https://app/after-auth")
+    );
+    let anotherPage = await browser.newPage();
+    await anotherPage.setRequestInterception(true);
+    anotherPage.on("request", async (request) => {
+      if (
+        request.url().includes("/samlproxy/sp/saml/sso") &&
+        request.method() == "GET"
+      ) {
+        await request.continue({
+          method: "POST",
+          postData: post_data,
+          headers: {
+            ...request.headers(),
+            "Content-Type": "application/x-www-form-urlencoded",
+          },
+        });
+      } else {
+        await request.continue();
+      }
+    });
+    await anotherPage.goto(`${saml_proxy_url}/samlproxy/sp/saml/sso`);
 
-  await page.goto(authorization_url);
+    // assertion
+    await isSensitiveError(anotherPage);
+  });
 
-  // assertion
-  await isSensitiveError(page);
-  await browser.close();
-});
+  test("Empty SSO", async () => {
+    const page = await browser.newPage();
+    await page.goto(`${saml_proxy_url}/samlproxy/idp/saml/sso`);
+    await page.waitForSelector(".usa-alert-error");
+    await isSensitiveError(page);
+  });
 
-test("404", async () => {
-  const browser = await puppeteer.launch(launchArgs);
-  const page = await browser.newPage();
+  test("404", async () => {
+    const page = await browser.newPage();
+    await page.goto(`${saml_proxy_url}/bad`);
+    await page.waitForSelector(".usa-alert-error");
+    await isError(page, "Error", "Route Not Found");
+  });
 
-  await page.goto("http://localhost:7000/bad");
+  test("modify", async () => {
+    const page = await browser.newPage();
+    await requestToken(page);
+    await authentication(page, "va.api.user+idme.001@gmail.com", true);
 
-  await browser.close();
+    page.on("request", async (request) => {
+      if (
+        request.url().includes("/samlproxy/sp/saml/sso") &&
+        request.method() == "POST"
+      ) {
+        let post_data = await decode(request);
+        post_data.SAMLResponse = await ModifyAttack(
+          post_data.SAMLResponse,
+          "uuid",
+          "modify"
+        );
+        await request.continue({
+          method: "POST",
+          postData: await encode(post_data),
+          headers: {
+            ...request.headers(),
+            "Content-Type": "application/x-www-form-urlencoded",
+          },
+        });
+      } else {
+        await request.continue();
+      }
+    });
+
+    await page.waitForSelector(".usa-alert-error");
+
+    await isSensitiveError(page);
+  });
 });
 
 const requestToken = async (page) => {
@@ -231,4 +229,56 @@ const decode = async (request) => {
   });
 
   return post_data;
+};
+
+const isError = async (page, header, errorMessage) => {
+  const heading = await page.$eval(
+    ".usa-alert-body > .usa-alert-heading",
+    (elem) => elem.innerText
+  );
+  expect(heading).toEqual(header);
+
+  const errorText = await page.$eval(
+    ".usa-alert-body > p",
+    (elem) => elem.innerText
+  );
+  expect(errorText).toEqual(errorMessage);
+
+  const errorIdHeadding = await page.$eval(
+    ".usa-alert-body > .usa-alert-heading:nth-of-type(2)",
+    (elem) => elem.innerText
+  );
+  expect(errorIdHeadding).toEqual("Error ID");
+
+  const errorId = await page.$eval(
+    ".usa-alert-body > pre",
+    (elem) => elem.innerText
+  );
+  expect(errorId);
+};
+
+const isSensitiveError = async (page) => {
+  const heading = await page.$eval(
+    ".usa-alert-body > .usa-alert-heading",
+    (elem) => elem.innerText
+  );
+  expect(heading).toEqual("Error");
+
+  const errorText = await page.$eval(
+    ".usa-alert-body > p",
+    (elem) => elem.innerText
+  );
+  expect(errorText).toEqual("Your request could not be processed.");
+
+  const errorIdHeadding = await page.$eval(
+    ".usa-alert-body > .usa-alert-heading:nth-of-type(2)",
+    (elem) => elem.innerText
+  );
+  expect(errorIdHeadding).toEqual("Error ID");
+
+  const errorId = await page.$eval(
+    ".usa-alert-body > pre",
+    (elem) => elem.innerText
+  );
+  expect(errorId);
 };
